LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/visual-scrollIntoView-001-expected.txt



FAIL Element.scrollIntoView scrolls visually assert_greater_than: expected a number greater than 1 but got 1
